backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==4.1.3
black==26.3.1
boto3==1.42.86
botocore==1.42.86
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
emergentintegrations==0.1.0
fastapi==0.110.1
fastuuid==0.14.0
filelock==3.25.2
flake8==7.3.0
frozenlist==1.8.0
fsspec==2026.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.2
google-api-python-client==2.194.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-genai==1.71.0
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.9.2
idna==3.11
importlib_metadata==9.0.0
iniconfig==2.3.0
isort==8.0.1
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jq==1.11.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
librt==0.8.1
litellm==1.80.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
mpmath==1.3.0
multidict==6.7.1
mypy==1.20.0
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.4
oauthlib==3.3.1
openai==1.99.9
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==3.0.2
passlib==1.7.4
pathspec==1.0.4
pillow==12.2.0
platformdirs==4.9.6
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.20.0
PyJWT==2.12.1
pymongo==4.5.0
pyparsing==3.3.2
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.24
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
reportlab==4.4.10
requests==2.33.1
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
s5cmd==0.2.0
scikit-learn==1.8.0
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
stripe==15.0.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0+cpu
torchvision==0.26.0+cpu
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.25.0
watchfiles==1.1.1
websockets==16.0
yarl==1.23.0
zipp==3.23.0
